Script for parsing wikipedia MCC page into the MCC db format
This script reads the English wiki page about mobile country codes
in edit mode: http://en.wikipedia.org/w/index.php?title=
Mobile_country_code&action=edit and prints out a mapping between
MCC & MNC codes and countries. The output format is

  <MCC> <MNC> <ISO3166 country code> <Name of Country>

For example

  123 45 AA Kingdom of Anonymous Aardwarks
  123 46 AA Kingdom of Anonymous Aardwarks
  456 78 RB Republic of Banana

Note that the script is not intended to be a part of the build,
rather a tool for making manual updates of the file data/MCC
easier.